Combined the ConferenceConfigWithEdges and ConferenceConfig helper classes for unit testing into one class that is based on using edges.
